Stop rounding Bulk quantities, and repair the rows already rounded
Bulk produce is sold in decimal kilograms and every layer above the database
knew it: CreateBulkOrderRequest.quantity is a Double, Bulk-orders.php reads it
with floatval() ("Can be decimal for kg"), and rider_dispatch.php goes out of
its way not to round it ("would show a rider 20 for a 20.5 kg load"). The
columns were INT(11), so none of it survived the write -- a 20.5 kg order
stored 21.

The data is recoverable because the total is computed from the untruncated
value and only the quantity is truncated, so the real figure is
total_price / discounted_price. The migration widens the three columns to
DECIMAL(12,3) first, prints what drifted, then repairs under guards: only
where the two disagree by more than 0.001 and less than a whole unit, which
skips any listing repriced after the order rather than guessing at it. Listing
stock is reported, never rewritten -- remaining_quantity is also moved by
cancellations and admin corrections, so the expected figure is a hint, not an
authority.

Verified by reproducing the bug: five orders at 20.5, 33.333, 100, 7.25 and
55.75 stored as 21/33/100/7/56, plus a sixth against a repriced listing. After
the migration all five read back exactly as ordered, the repriced one is
untouched and flagged for a human, and a second run is a clean no-op.

Code that had to follow:

  admin/Bulk-listings.php cast quantities to int in five places. The damaging
  one computed the stock delta from a truncated original, so correcting a
  20.5 kg listing shifted remaining stock by the wrong amount. Its float
  comparisons now use a tolerance, since === on floats reports changes that
  did not happen.

  The Kotlin models were Int, and Gson silently coerced 20.500 to 20 -- the
  app showed a different quantity from the one the server held. Now Double,
  with formatQuantity() so a whole 100.0 still reads "100" rather than "100.0".

  The vendor form's quantity fields filtered input to digits only, which made
  a decimal impossible to type: 20.5 became 205. decimalInput() keeps one
  separator, normalises a comma, and caps at the column's three places.

The contract test caught the discount floor change on its own, which is what
it is for. Note one pre-existing failure in RoutePolylineMathTest, unrelated
to any of this and confirmed failing on a clean tree.

+++ app/src/main/java/com/techaus/afamfresh/models/SurplusModels.kt
@@ -25,8 +25,12 @@ data class BulkListing(
     @SerializedName("discount_percent") val discountPercent: Double = 0.0,
     @SerializedName("discounted_price") val discountedPrice: Double = 0.0,
 
-    @SerializedName("Bulk_quantity") val BulkQuantity: Int = 0,
-    @SerializedName("remaining_quantity") val remainingQuantity: Int = 0,
+    // Double, not Int. Bulk produce is sold in decimal kilograms, and the
+    // columns are DECIMAL(12,3) as of the decimal-quantities migration. Gson
+    // coerced 20.500 into an Int silently, so the app displayed a different
+    // quantity from the one the server held.
+    @SerializedName("Bulk_quantity") val BulkQuantity: Double = 0.0,
+    @SerializedName("remaining_quantity") val remainingQuantity: Double = 0.0,
 
     /** datetime, e.g. "2026-08-04 18:00:00" — NOT an ISO-8601 instant. */
     @SerializedName("expiry_date") val expiryDate: String? = null,
@@ -68,7 +72,7 @@ data class BulkListing(
             ?: listOfNotNull(vendorFirstName, vendorLastName)
                 .joinToString(" ").trim().takeIf { it.isNotBlank() }
 
-    val isSoldOut: Boolean get() = remainingQuantity <= 0
+    val isSoldOut: Boolean get() = remainingQuantity <= 0.0
     val isApproved: Boolean get() = status == "approved"
 }
 
@@ -101,7 +105,7 @@ data class CreateBulkListingRequest(
     @SerializedName("product_id") val productId: Int,
     @SerializedName("original_price") val originalPrice: Double,
     @SerializedName("discount_percent") val discountPercent: Double,
-    @SerializedName("Bulk_quantity") val BulkQuantity: Int,
+    @SerializedName("Bulk_quantity") val BulkQuantity: Double,
     /** "YYYY-MM-DD HH:MM:SS" */
     @SerializedName("expiry_date") val expiryDate: String,
     @SerializedName("listing_type") val listingType: String = "goodie_bag",
@@ -136,13 +140,14 @@ data class CreateBulkListingRequest(
  * cannot edit price, quantity or expiry, which is what the app's old
  * `updateVendorBulkListing` tried to send.
  *
- * ⚠️ SECURITY: the PHP performs no ownership check, so any authenticated caller
- * can set any listing's status — including approving their own. This is
- * reported to the backend owner; the app only ever sends `remaining_quantity`.
+ * The ownership hole this used to warn about is closed: the endpoint now
+ * resolves the listing's owner and refuses a caller who is neither that vendor
+ * nor an admin, and status/admin_notes are admin-only. The app still sends
+ * only `remaining_quantity`.
  */
 data class UpdateBulkListingRequest(
     @SerializedName("listing_id") val listingId: Int,
-    @SerializedName("remaining_quantity") val remainingQuantity: Int? = null,
+    @SerializedName("remaining_quantity") val remainingQuantity: Double? = null,
     @SerializedName("status") val status: String? = null,
     @SerializedName("admin_notes") val adminNotes: String? = null
 )
