+++ app/src/main/java/com/techaus/afamfresh/repository/VendorRepository.kt
@@ -236,7 +236,7 @@ class VendorRepository(
      */
     fun updateListingQuantity(
         listingId: Int,
-        remainingQuantity: Int,
+        remainingQuantity: Double,
         callback: (Boolean, ApiError?) -> Unit
     ) {
         apiService.updateVendorBulkListing(
